refactor: move backend-specific machinery to own ts module
